v0.2.0 :: use healthchecks instead of homebrew server
This commit is a major overhaul of runner-rs that converts it to use
healthchecks.io instead of my previous homebrew HomeAssistant webhooks.

As part of that, I've converted it to use tokio and tracing (mostly for
practice), updated several dependencies, and tests.

If this works well for me, I'll likely merge this to master in the
coming weeks.